chrome->browser and fix manifest
remove entries in manifest that are not supported
oauth is not supported and is only used for GDrive - which wasn't working anyway. TODO: remove the Gdrive option from the UI

firefox javascript api are under firefox instead of chrome

rename refferences to chrome-extension to moz-extension

pass appUrl in callback. Firefox seems to handle garbage collection differently to chrome